Write PDFs
This brings in the itext library for writing PDFs. Still some issues
though.

It appears as though itext only reads the unevaluated excel. Could be
just the way I'm using it though; maybe the classes workbook is not
evaluated by evaluateAll maybe a copy of it is evaluated instead. In
that case it may be as easy as just calling evaluateAll inside the toPdf
function.