Update web app styling and add resource files.
This is a complete change of the web app. I've switched from bootstrap and vanilla JS to Materialize and mostly
Jquery. The functionality is still the same, but the look is much different. We now have a static side
navigation bar for large screens and a collapsable side navigation bar for medium and smaller screens.